Prevent empty portal pane item from being serialized or copied
